Update model architectures and data handling for training
The commit introduces major refinements in the training script, especially concerning model architectures and data processing. It expands the choice of neural network architectures by including options for ConvNeXt and its pretrained version. Specific handling for these models is included in the training setup, like the loading of pretrained weights. The data loading process during training is modified to use testing data instead of validation data. Minor adaptations are made in other areas like the loss function and environment setup for distributed processing. More configurations are allowed through changes in the YAML file. Also, there is an addition of a new model file for the pretrained ConvNeXt.